+++ src/logistics/Energetics.ts
@@ -20,9 +20,9 @@ export class Energetics {
 			},
 			energy: {
 				sendSize    : 25000,	// Send energy in chunks of this size
-				inThreshold : 50000, 	// Terminals with < this amount of energy in room actively store energy
-				outThreshold: 150000,	// Terminals with more than this amount of energy in store send elsewhere
-				equilibrium : 100000, 	// Try to maintain this amount; should be energyInThreshold + 2*energySendSize
+				inThreshold : 25000, 	// Terminals with < this amount of energy in room actively store energy
+				outThreshold: 100000,	// Terminals with more than this amount of energy in store send elsewhere
+				equilibrium : 50000, 	// Try to maintain this amount; should be energyInThreshold + 2*energySendSize
 				tolerance   : 5000,		// Don't care about deviation by less than this amount
 			},
 		},